コンポーネントの分け方再考
