+++ app/src/app/(main)/_components/avatar-and-user-menu/index.ts
@@ -1,0 +1,1 @@
+export * from './avatar-and-user-menu';
