+++ app/src/app/(main)/_components/header/header.css.ts
@@ -21,59 +21,3 @@ export const link = style({
   textDecoration: 'none',
   color: theme.color.token.semantic.text,
 });
-
-export const userAvatar = style({
-  borderRadius: '50%',
-  cursor: 'pointer',
-});
-
-export const popover = style({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'space-between',
-  padding: '8px',
-  borderRadius: theme.size.radius.normal,
-  backgroundColor: theme.color.token.semantic.background,
-  height: '120px',
-  width: '200px',
-});
-
-export const accountInfoArea = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '4px',
-});
-
-export const accountName = style({
-  margin: 0,
-  fontWeight: 'bold',
-});
-
-export const displayName = style({
-  margin: 0,
-  color: theme.color.token.semantic.textMuted,
-});
-
-export const popoverButtonArea = style({
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'flex-end',
-});
-
-export const dialog = style({
-  width: '400px',
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '8px',
-  padding: '8px',
-});
-
-export const dialogButtonArea = style({
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'center',
-});
-
-export const dialogAnnotation = style({
-  color: theme.color.token.semantic.textMuted,
-});
