+++ app/src/app/(main)/_components/login-button-and-dialog/login-button-and-dialog.stories.tsx
@@ -1,0 +1,13 @@
+import { LoginButtonAndDialog } from './login-button-and-dialog';
+
+import type { Meta, StoryObj } from '@storybook/react';
+
+export default {
+  component: LoginButtonAndDialog,
+} as Meta<typeof LoginButtonAndDialog>;
+
+type Story = StoryObj<typeof LoginButtonAndDialog>;
+
+export const Default = {
+  args: {},
+} satisfies Story;
